address comments also rename Implicits
