+++ core/src/test/scala/com/krux/hyperion/workflow/WorkflowExpressionSpec.scala
@@ -3,7 +3,7 @@ package com.krux.hyperion.workflow
 import com.krux.hyperion.activity.ShellCommandActivity
 import com.krux.hyperion.HyperionContext
 import com.krux.hyperion.resource.Ec2Resource
-import com.krux.hyperion.workflow.Implicits._
+import com.krux.hyperion.workflow.WorkflowExpression._
 import com.typesafe.config.ConfigFactory
 import org.scalatest.WordSpec
